style: move in-flight exits to its test module
